# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OriCMS API
  version: 2.0.0
  description: |
    Current OpenAPI snapshot for the main live OriCMS route families.

    This spec is intentionally curated rather than exhaustive. It focuses on the
    active project-based API model and the current agent gateway, and it avoids
    the older site-based and tier-era contracts that used to live in this file.

    Canonical routing model:
    - management API: `/api/v1/*`
    - project-scoped routes: `/api/v1/projects/:projectId/*`
    - delivery routes: `/api/v1/delivery/projects/:projectId/*`
    - agent gateway: `/api/v1/agent/v1/*`
servers:
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Log out
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/projects/{projectId}/collections/{collectionId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/CollectionId'
    get:
      tags: [Collections]
      summary: List entries in a collection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Entries returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
    post:
      tags: [Collections]
      summary: Create an entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
    delete:
      tags: [Collections]
      summary: Delete a collection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Collection deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
  /api/v1/projects/{projectId}/collections/{collectionId}/{entryId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/CollectionId'
      - $ref: '#/components/parameters/EntryId'
    get:
      tags: [Collections]
      summary: Get an entry
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Entry returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
    put:
      tags: [Collections]
      summary: Update an entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
        '409':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Collections]
      summary: Delete an entry
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Entry deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: AgentToken
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CollectionId:
      name: collectionId
      in: path
      required: true
      schema:
        type: string
...
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: string
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          additionalProperties: true
        meta:
          type: object
          additionalProperties: true
    GenericSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
    AuthSuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
            refreshToken:
              type: string
...